portaudio_sys

Function Pa_OpenDefaultStream

pub unsafe extern "C" fn Pa_OpenDefaultStream(
    stream: *mut *mut PaStream,
    numInputChannels: c_int,
    numOutputChannels: c_int,
    sampleFormat: PaSampleFormat,
    sampleRate: f64,
    framesPerBuffer: c_ulong,
    streamCallback: PaStreamCallback,
    userData: *mut c_void,
) -> PaError

A simplified version of Pa_OpenStream() that opens the default input and/or output devices.

@param stream The address of a PaStream pointer which will receive a pointer to the newly opened stream.

@param numInputChannels The number of channels of sound that will be supplied to the stream callback or returned by Pa_ReadStream. It can range from 1 to the value of maxInputChannels in the PaDeviceInfo record for the default input device. If 0 the stream is opened as an output-only stream.

@param numOutputChannels The number of channels of sound to be delivered to the stream callback or passed to Pa_WriteStream. It can range from 1 to the value of maxOutputChannels in the PaDeviceInfo record for the default output device. If 0 the stream is opened as an output-only stream.

@param sampleFormat The sample format of both the input and output buffers provided to the callback or passed to and from Pa_ReadStream and Pa_WriteStream. sampleFormat may be any of the formats described by the PaSampleFormat enumeration.

@param sampleRate Same as Pa_OpenStream parameter of the same name. @param framesPerBuffer Same as Pa_OpenStream parameter of the same name. @param streamCallback Same as Pa_OpenStream parameter of the same name. @param userData Same as Pa_OpenStream parameter of the same name.

@return As for Pa_OpenStream

@see Pa_OpenStream, PaStreamCallback