portaudio_sys

Function Pa_SetStreamFinishedCallback

pub unsafe extern "C" fn Pa_SetStreamFinishedCallback(
    stream: *mut PaStream,
    streamFinishedCallback: PaStreamFinishedCallback,
) -> PaError

Register a stream finished callback function which will be called when the stream becomes inactive. See the description of PaStreamFinishedCallback for further details about when the callback will be called.

@param stream a pointer to a PaStream that is in the stopped state - if the stream is not stopped, the stream’s finished callback will remain unchanged and an error code will be returned.

@param streamFinishedCallback a pointer to a function with the same signature as PaStreamFinishedCallback, that will be called when the stream becomes inactive. Passing NULL for this parameter will un-register a previously registered stream finished callback function.

@return on success returns paNoError, otherwise an error code indicating the cause of the error.

@see PaStreamFinishedCallback