portaudio_sys

Type Alias PaStreamCallback

pub type PaStreamCallback = Option<unsafe extern "C" fn(input: *const c_void, output: *mut c_void, frameCount: c_ulong, timeInfo: *const PaStreamCallbackTimeInfo, statusFlags: PaStreamCallbackFlags, userData: *mut c_void) -> c_int>;

Functions of type PaStreamCallback are implemented by PortAudio clients. They consume, process or generate audio in response to requests from an active PortAudio stream.

When a stream is running, PortAudio calls the stream callback periodically. The callback function is responsible for processing buffers of audio samples passed via the input and output parameters.

The PortAudio stream callback runs at very high or real-time priority. It is required to consistently meet its time deadlines. Do not allocate memory, access the file system, call library functions or call other functions from the stream callback that may block or take an unpredictable amount of time to complete.

In order for a stream to maintain glitch-free operation the callback must consume and return audio data faster than it is recorded and/or played. PortAudio anticipates that each callback invocation may execute for a duration approaching the duration of frameCount audio frames at the stream sample rate. It is reasonable to expect to be able to utilise 70% or more of the available CPU time in the PortAudio callback. However, due to buffer size adaption and other factors, not all host APIs are able to guarantee audio stability under heavy CPU load with arbitrary fixed callback buffer sizes. When high callback CPU utilisation is required the most robust behavior can be achieved by using paFramesPerBufferUnspecified as the Pa_OpenStream() framesPerBuffer parameter.

@param input and @param output are either arrays of interleaved samples or; if non-interleaved samples were requested using the paNonInterleaved sample format flag, an array of buffer pointers, one non-interleaved buffer for each channel.

The format, packing and number of channels used by the buffers are determined by parameters to Pa_OpenStream().

@param frameCount The number of sample frames to be processed by the stream callback.

@param timeInfo Timestamps indicating the ADC capture time of the first sample in the input buffer, the DAC output time of the first sample in the output buffer and the time the callback was invoked. See PaStreamCallbackTimeInfo and Pa_GetStreamTime()

@param statusFlags Flags indicating whether input and/or output buffers have been inserted or will be dropped to overcome underflow or overflow conditions.

@param userData The value of a user supplied pointer passed to Pa_OpenStream() intended for storing synthesis data etc.

@return The stream callback should return one of the values in the ::PaStreamCallbackResult enumeration. To ensure that the callback continues to be called, it should return paContinue (0). Either paComplete or paAbort can be returned to finish stream processing, after either of these values is returned the callback will not be called again. If paAbort is returned the stream will finish as soon as possible. If paComplete is returned, the stream will continue until all buffers generated by the callback have been played. This may be useful in applications such as soundfile players where a specific duration of output is required. However, it is not necessary to utilize this mechanism as Pa_StopStream(), Pa_AbortStream() or Pa_CloseStream() can also be used to stop the stream. The callback must always fill the entire output buffer irrespective of its return value.

@see Pa_OpenStream, Pa_OpenDefaultStream

@note With the exception of Pa_GetStreamCpuLoad() it is not permissible to call PortAudio API functions from within the stream callback.

Aliased Type

enum PaStreamCallback {
    None,
    Some(unsafe extern "C" fn(_: *const c_void, _: *mut c_void, _: u64, _: *const PaStreamCallbackTimeInfo, _: u64, _: *mut c_void) -> i32),
}

Variants

None

No value.

Some(unsafe extern "C" fn(_: *const c_void, _: *mut c_void, _: u64, _: *const PaStreamCallbackTimeInfo, _: u64, _: *mut c_void) -> i32)

Some value of type T.